proper packaging + RowEncoders + implementation of iterrows
