+++ core/src/main/scala/org/sciscala/simpledf/Encoder.scala
