+++ core/src/main/scala/org/sciscala/simpledf/package.scala
